Moving classes and structures in subfolders and adapting namespaces
In the project Core there was only one namespace, which made it impossible to get an overview of the containing classes. For this reason, subfolders have been added and the classes moved in as far as possible. Furthermore, the namespaces of the files were revised and adapted to the rules of .NET.